redacted string tests
